+++ packages/mg-components/.storybook/main.ts
@@ -38,7 +38,6 @@ const config: StorybookConfig = {
   },
   addons: [
     getAbsolutePath('@storybook/addon-essentials'),
-    getAbsolutePath('@pxtrn/storybook-addon-docs-stencil'),
     getAbsolutePath('@storybook/addon-a11y'),
     getAbsolutePath('@storybook/addon-links'),
     getAbsolutePath('@storybook/addon-interactions'),
@@ -120,6 +119,9 @@ const config: StorybookConfig = {
         'v5.27.1': 'https://626149b307606d003ada26b4-jqeifeltaz.chromatic.com',
         'v5.27.2': 'https://626149b307606d003ada26b4-rgtwzbiszp.chromatic.com',
         'v5.27.3': 'https://626149b307606d003ada26b4-nxqzxvfuib.chromatic.com',
+        'v5.27.4': 'https://626149b307606d003ada26b4-pponmjenit.chromatic.com',
+        'v5.28.0': 'https://626149b307606d003ada26b4-grmjndcjgx.chromatic.com',
+        'v5.28.1': 'https://626149b307606d003ada26b4-vgqkjpyvnp.chromatic.com',
       },
     },
   },