+++ packages/mg-components/src/components/atoms/mg-tooltip/doc/mg-tooltip.stories.tsx
@@ -8,11 +8,6 @@ export default {
   parameters: {
     layout: 'centered',
   },
-  argTypes: {
-    placement: {
-      control: { type: 'select' },
-    },
-  },
 };
 
 /**