+++ packages/mg-components/src/components/molecules/inputs/mg-input-password/test/mg-input-password.spec.tsx
@@ -34,6 +34,7 @@ describe('mg-input-password', () => {
     { label: 'label', identifier: 'identifier', tooltip: 'My Tooltip Message', labelOnTop: true },
     { label: 'label', identifier: 'identifier', tooltip: 'My Tooltip Message', tooltipPosition: 'label' },
     { label: 'label', identifier: 'identifier', tooltip: 'My Tooltip Message', tooltipPosition: 'input', labelOnTop: true },
+    { label: 'label', identifier: 'identifier', maxlength: 20 },
   ])('Should render with args %s:', async args => {
     const { root } = await getPage(args);
     expect(root).toMatchSnapshot();