+++ packages/mg-components/stencil.config.ts
@@ -2,6 +2,7 @@ import { webTypesGenerator, vsCodeGenerator } from '@mgdis/stencil-helpers';
 import { writeFile, mkdir } from 'node:fs/promises';
 import { dirname } from 'node:path';
 import { Config } from '@stencil/core';
+import { coverageReporters, coverageThreshold } from 'jest-config/base';
 import packageJson from './package.json';
 const { name, version, 'web-types': webTypes, contributes } = packageJson;
 
@@ -45,7 +46,7 @@ export const config: Config = {
     },
     {
       type: 'docs-json',
-      file: '.storybook/docs/components.json',
+      file: '.storybook/components.json',
     },
     {
       type: 'docs-readme',
@@ -92,6 +93,7 @@ export const config: Config = {
   },
   testing: {
     setupFilesAfterEnv: ['./jest.setup.ts'],
-    coverageReporters: ['cobertura', 'lcov', 'html', 'text'],
+    coverageReporters,
+    coverageThreshold,
   },
 };