+++ scripts/run-e2e.sh
@@ -28,6 +28,7 @@ pnpm build "$@"
 # Prepare root package.json
 prepare_package "package.json" '{
   name, 
+  packageManager,
   "scripts": {
     "apps:notification-center": .scripts."apps:notification-center",
     "test:e2e": .scripts."test:e2e"
@@ -105,7 +106,8 @@ prepare_package "packages/stencil-helpers/package.json" '{
 # Create turbo.json
 echo '{
   "$schema": "https://turborepo.org/schema.json",
-  "pipeline": {
+  "globalEnv": ["PLAYWRIGHT_BROWSERS_PATH"],
+  "tasks": {
     "test:e2e": {}
   }
 }' > "$temp_dir/turbo.json"