+++ packages/eslint-config-vue3/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 2.0.8
+
+### Patch Changes
+
+- 9c9620c: update dependencies
+- Updated dependencies [9c9620c]
+  - @mgdis/eslint-plugin-vue@4.0.6
+
+## 2.0.7
+
+### Patch Changes
+
+- c567b43: update dependencies
+- Updated dependencies [c567b43]
+  - @mgdis/eslint-plugin-vue@4.0.5
+
 ## 2.0.6
 
 ### Patch Changes