+++ packages/eslint-plugin-vue/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 4.0.6
+
+### Patch Changes
+
+- 9c9620c: update dependencies
+
+## 4.0.5
+
+### Patch Changes
+
+- c567b43: update dependencies
+
 ## 4.0.4
 
 ### Patch Changes