+++ packages/playwright-helpers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mgdis/playwright-helpers
 
+## 1.0.9
+
+### Patch Changes
+
+- 6195e7a: Update PW
+
+## 1.0.8
+
+### Patch Changes
+
+- c567b43: update dependencies
+
 ## 1.0.7
 
 ### Patch Changes