+++ packages/stencil-helpers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @mgdis/stencil-helpers
 
+## 1.1.8
+
+### Patch Changes
+
+- 9c9620c: update dependencies
+
+## 1.1.7
+
+### Patch Changes
+
+- 4ee49fe: Update dependencies
+
+## 1.1.6
+
+### Patch Changes
+
+- c567b43: update dependencies
+
 ## 1.1.5
 
 ### Patch Changes